Fix warnings.

+++ src/text2code.cpp
@@ -8,11 +8,11 @@ text2code(text_t *text, cmd_arr_t *cmd_arr)
         //fprintf(stderr, "num_of_lines: %zu\n", text->num_of_lines);
         int *cmd_array = (int *) calloc(text->num_of_lines * 2, sizeof(int));
 
-        int cmd_count = 0;
-
+        size_t  cmd_count = 0;
         char cmd_name[MAX_CMD_SIZE] = {};
         int val = 0;
         size_t line_count = 0;
+
         while (line_count < text->num_of_lines) {
                 sscanf(text->lines[line_count].first_ch, "%s", cmd_name);
